+++ NEWS.md
@@ -2,6 +2,9 @@
 
 * Vectorize `igo_members()`
 * Generalize search ignoring case.
+* Add `igo_state_membership()`
+* `cow_country_codes` now is internal. 
+* Added `igoR::states2016` in replacement of `cow_country_codes`.
 
 
 # igoR 0.0.0.9000